task2_task3/requirements.txt
requests
pandas
python-dotenv
pytest
pylint
streamlit
altair
vl-convert-python